データベースにアクセスする処理をaccessDB.tsに共通化
